Extract Element protocols into it's own file
